// generated/api/models/schedulePaymentStatus.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * WolfBankGateway
 * OpenAPI spec version: 1.0
 */

export type SchedulePaymentStatus =
  (typeof SchedulePaymentStatus)[keyof typeof SchedulePaymentStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SchedulePaymentStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2
} as const;